+++ library/js/customize-control-multicheckbox.js
@@ -7,11 +7,10 @@
  * @subpackage  Customize
  *
  * @since    2.1.0
- * @version  2.1.1
+ * @version  2.2.0
  */
 ( function( exports, $ ) {
-
-	$( document ).ready( function() {
+	$( wp.customize ).on( 'ready', function() {
 
 
 
@@ -52,5 +51,4 @@
 
 
 	} );
-
 } )( wp, jQuery );